fix(ci): repair PC Controller presenter firmware compile errors
ESP-IDF v5.5.3 / GCC 14 toolchain uplift 暴露了一组 presenter 档
main/ 模块的 API/语法漂移。workflow 已能正常进入 build 步骤
(action 1/2 均成功),但 ninja 在编译 main/ 源时全部 FAILED:

  pc_storage.c:208        nvs_delete_key 已重命名为 nvs_erase_key
                          (implicit-function-declaration)
  pc_app_main.c:84-88     跨匿名枚举 _Static_assert 触发
                          -Werror=enum-compare,显式 int 转换规避
  pc_ble_hid.c:180        NimBLE 新版 ble_hs_mbuf_to_flat 签名
                          收紧为 4 参数(uint16_t max_len,
                          uint16_t *out_copy_len),旧 5 参数调用
                          触发 too many arguments/-Wint-conversion
  pc_ble_hid.c:412        ble_hs_adv_fields.appearance_is_complete
                          -> appearance_is_present(API 字段名变更)
  pc_ble_hid.c:765,787    BLE_ERR_REM_USER_CONN_TERMINATED ->
                          BLE_ERR_REM_USER_CONN_TERM
  sdkconfig.defaults      presenter 档启用 LV_FONT_UNSCII_8
                          (8px placeholder HUD 字体此前未开)
  pc_ui_fui*.c            LVGL v9 移除 lv_obj_set_style_text_shadow_*,
                          文本发光效果改走 box shadow
                          (shadow_color/width/opa 保持视觉一致)

参考本地抓取的 CI 日志:
  workflow-job-run-cdc03933-1a8b-5479-8632-25fe80c4b449

+++ main/pc_app_main.c
@@ -80,12 +80,14 @@ static const char *TAG = "pc_app";
 #define FX_CAP 4                /* FSM 单次最大 3 个 effect,留 1 个余量 */
 
 /* 纯逻辑枚举与 BSP 枚举逐值同构校验(规格 §5:按键事件数值透传,
- * 无需转换)。任一侧改动都会在此编译期炸出,防止静默错位。 */
-_Static_assert(PC_BTN_UP == BSP_BTN_UP && PC_BTN_DOWN == BSP_BTN_DOWN &&
-                   PC_BTN_OK == BSP_BTN_OK,
+ * 无需转换)。任一侧改动都会在此编译期炸出,防止静默错位。
+ * 注:GCC 升级后 -Werror=enum-compare 会拒绝跨匿名枚举比较,
+ * 故显式转 int 再比较。 */
+_Static_assert((int)PC_BTN_UP == (int)BSP_BTN_UP && (int)PC_BTN_DOWN == (int)BSP_BTN_DOWN &&
+                   (int)PC_BTN_OK == (int)BSP_BTN_OK,
                "pc_btn_t / bsp_btn_t enumeration mismatch");
-_Static_assert(PC_EV_PRESS == BSP_BTN_PRESS && PC_EV_CLICK == BSP_BTN_CLICK &&
-                   PC_EV_DOUBLE == BSP_BTN_DOUBLE && PC_EV_LONG == BSP_BTN_LONG,
+_Static_assert((int)PC_EV_PRESS == (int)BSP_BTN_PRESS && (int)PC_EV_CLICK == (int)BSP_BTN_CLICK &&
+                   (int)PC_EV_DOUBLE == (int)BSP_BTN_DOUBLE && (int)PC_EV_LONG == (int)BSP_BTN_LONG,
                "pc_btn_ev_t / bsp_btn_ev_t enumeration mismatch");
 
 /* ---- 统一事件队列(规格 §7 行 162) ---- */
